% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{CircularPipe}
\alias{CircularPipe}
\title{Pipe class with circular cross-section}
\description{
A `CircularPipe` is created by a diameter and slope. Units passed to the
`depth` argument of methods must match the pipe units
}
\section{Super class}{
\code{bes.hydro::Circle} -> \code{CircularPipe}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{diameter}}{the pipe diameter}

\item{\code{n}}{a default roughness coefficient for the pipe}

\item{\code{slope}}{the pipe slope}

\item{\code{unit}}{the pipe diameter unit}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CircularPipe-new}{\code{CircularPipe$new()}}
\item \href{#method-CircularPipe-print}{\code{CircularPipe$print()}}
\item \href{#method-CircularPipe-hydraulic_radius}{\code{CircularPipe$hydraulic_radius()}}
\item \href{#method-CircularPipe-wetted_perimeter}{\code{CircularPipe$wetted_perimeter()}}
\item \href{#method-CircularPipe-wetted_area}{\code{CircularPipe$wetted_area()}}
\item \href{#method-CircularPipe-critical_flow}{\code{CircularPipe$critical_flow()}}
\item \href{#method-CircularPipe-convert_to_si}{\code{CircularPipe$convert_to_si()}}
\item \href{#method-CircularPipe-clone}{\code{CircularPipe$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bes.hydro" data-topic="Circle" data-id="coordinates"><a href='../../bes.hydro/html/Circle.html#method-Circle-coordinates'><code>bes.hydro::Circle$coordinates()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bes.hydro" data-topic="Circle" data-id="segment_arc_length"><a href='../../bes.hydro/html/Circle.html#method-Circle-segment_arc_length'><code>bes.hydro::Circle$segment_arc_length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bes.hydro" data-topic="Circle" data-id="segment_area"><a href='../../bes.hydro/html/Circle.html#method-Circle-segment_area'><code>bes.hydro::Circle$segment_area()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bes.hydro" data-topic="Circle" data-id="segment_central_angle"><a href='../../bes.hydro/html/Circle.html#method-Circle-segment_central_angle'><code>bes.hydro::Circle$segment_central_angle()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CircularPipe-new"></a>}}
\if{latex}{\out{\hypertarget{method-CircularPipe-new}{}}}
\subsection{Method \code{new()}}{
Create a new `CircularPipe` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CircularPipe$new(diameter, slope, n = NA, unit = c("ft", "in", "m"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{diameter}}{the pipe diameter}

\item{\code{slope}}{the pipe slope}

\item{\code{n}}{a default Manning roughness coefficient}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new `CircularPipe` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CircularPipe-print"></a>}}
\if{latex}{\out{\hypertarget{method-CircularPipe-print}{}}}
\subsection{Method \code{print()}}{
Print a CircularPipe object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CircularPipe$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CircularPipe-hydraulic_radius"></a>}}
\if{latex}{\out{\hypertarget{method-CircularPipe-hydraulic_radius}{}}}
\subsection{Method \code{hydraulic_radius()}}{
Calculate the pipe hydraulic radius
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CircularPipe$hydraulic_radius(depth)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{depth}}{depth of flow}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CircularPipe-wetted_perimeter"></a>}}
\if{latex}{\out{\hypertarget{method-CircularPipe-wetted_perimeter}{}}}
\subsection{Method \code{wetted_perimeter()}}{
Calculate the pipe wetted perimeter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CircularPipe$wetted_perimeter(depth)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{depth}}{depth of flow}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CircularPipe-wetted_area"></a>}}
\if{latex}{\out{\hypertarget{method-CircularPipe-wetted_area}{}}}
\subsection{Method \code{wetted_area()}}{
Calculate the pipe wetted area
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CircularPipe$wetted_area(depth)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{depth}}{depth of flow}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CircularPipe-critical_flow"></a>}}
\if{latex}{\out{\hypertarget{method-CircularPipe-critical_flow}{}}}
\subsection{Method \code{critical_flow()}}{
Calculate critical flow. See Chow "Open Channel Hydraulics" Page 81
Equation 4-21
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CircularPipe$critical_flow(depth)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{depth}}{depth of flow}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CircularPipe-convert_to_si"></a>}}
\if{latex}{\out{\hypertarget{method-CircularPipe-convert_to_si}{}}}
\subsection{Method \code{convert_to_si()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CircularPipe$convert_to_si()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CircularPipe-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CircularPipe-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CircularPipe$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
