% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manning.R
\name{manning}
\alias{manning}
\title{Calculate velocity estimated by Manning's equation}
\arguments{
\item{pipe}{a pipe geometry object}

\item{depth}{depth measurements}

\item{n}{Manning's roughness coefficient}

\item{unit}{the unit of measure for depth measurements, defaults to feet.}
}
\description{
Calculate flow or velocity using velocity estimated by Manning's equation and pipe
geometry. If missing, will use the default roughness defined for the pipe,
if available.
}
