% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_circular.R
\name{PolygonPipe}
\alias{PolygonPipe}
\title{Pipe class with arbitrary polygon cross-section}
\description{
A PolygonPipe is created by passing a shape of class `sfg` from the sf
package. Units passed to the `depth` argument of methods must match the pipe
units
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{shape}}{the pipe cross-section}

\item{\code{n}}{a default roughness coefficient for the pipe}

\item{\code{slope}}{the pipe slope}

\item{\code{bbox}}{the bounding box of the pipe shape}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PolygonPipe-new}{\code{PolygonPipe$new()}}
\item \href{#method-PolygonPipe-print}{\code{PolygonPipe$print()}}
\item \href{#method-PolygonPipe-wetted_area}{\code{PolygonPipe$wetted_area()}}
\item \href{#method-PolygonPipe-wetted_perimeter}{\code{PolygonPipe$wetted_perimeter()}}
\item \href{#method-PolygonPipe-hydraulic_radius}{\code{PolygonPipe$hydraulic_radius()}}
\item \href{#method-PolygonPipe-clone}{\code{PolygonPipe$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolygonPipe-new"></a>}}
\if{latex}{\out{\hypertarget{method-PolygonPipe-new}{}}}
\subsection{Method \code{new()}}{
Create a new `PolygonPipe` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolygonPipe$new(shape, slope, n = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{shape}}{the pipe cross-section, an `sfg` object}

\item{\code{slope}}{the pipe slope}

\item{\code{n}}{a default Manning roughness coefficient}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new `PolygonPipe` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolygonPipe-print"></a>}}
\if{latex}{\out{\hypertarget{method-PolygonPipe-print}{}}}
\subsection{Method \code{print()}}{
Print a PolygonPipe object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolygonPipe$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolygonPipe-wetted_area"></a>}}
\if{latex}{\out{\hypertarget{method-PolygonPipe-wetted_area}{}}}
\subsection{Method \code{wetted_area()}}{
The pipe wetted area
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolygonPipe$wetted_area(depth, interp = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{depth}}{depth of flow}

\item{\code{interp}}{TRUE if desired to interpolate pipe geometry calculations
rather than calculate exact values. May speed up calculations for large
depth vectors}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector of wetted areas for the given depth
Calculate pipe wetted perimeter
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolygonPipe-wetted_perimeter"></a>}}
\if{latex}{\out{\hypertarget{method-PolygonPipe-wetted_perimeter}{}}}
\subsection{Method \code{wetted_perimeter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolygonPipe$wetted_perimeter(depth, interp = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{depth}}{depth of flow}

\item{\code{interp}}{TRUE if desired to interpolate pipe geometry calculations
rather than calculate exact values. May speed up calculations for large
depth vectors}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector of wetted perimeters for the given depth
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolygonPipe-hydraulic_radius"></a>}}
\if{latex}{\out{\hypertarget{method-PolygonPipe-hydraulic_radius}{}}}
\subsection{Method \code{hydraulic_radius()}}{
Calculate hydraulic radius for the given depth vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolygonPipe$hydraulic_radius(depth, interp = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{depth}}{depth of flow}

\item{\code{interp}}{TRUE if desired to interpolate pipe geometry calculations
rather than calculate exact values. May speed up calculations for large
depth vectors}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector of wetted perimeters for the given depth
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolygonPipe-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PolygonPipe-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolygonPipe$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
