% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro.R, R/manning.R
\name{var_n_factor}
\alias{var_n_factor}
\title{Wong, Tommy SW. "Discussion of" Explicit solutions of the Manning equation
for partially filled circular pipes"(1)." Canadian Journal of Civil
Engineering 33.3 (2006): 349-353.}
\usage{
var_n_factor(theta)

var_n_factor(theta)
}
\description{
Wong, Tommy SW. "Discussion of" Explicit solutions of the Manning equation
for partially filled circular pipes"(1)." Canadian Journal of Civil
Engineering 33.3 (2006): 349-353.

Wong, Tommy SW. "Discussion of" Explicit solutions of the Manning equation
for partially filled circular pipes"(1)." Canadian Journal of Civil
Engineering 33.3 (2006): 349-353.
}
