% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_density.R
\name{water_density}
\alias{water_density}
\title{Calculate water density from pressure}
\usage{
water_density(x)
}
\arguments{
\item{x}{water temperature in Celsius}
}
\value{
a numeric vector, representing the density of water in kg/m^3
}
\description{
Calculates water density according to "The international association for the
properties of water and steam." Revised supplementary release on properties of
liquid water at 0.1 MPa (2011)
}
