% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_density.R
\name{water_depth}
\alias{water_depth}
\title{Calculates depth from water and air pressure}
\usage{
water_depth(water, air, temp)
}
\arguments{
\item{water}{water pressure in kPa}

\item{air}{air pressure in kPa}

\item{temp}{water temperature in °C}
}
\value{
numeric vector, representing depth in meters
}
\description{
Calculates depth from water and air pressure
}
